Pow(x, n): Refactored solution resulted in same performance

+++ submissions/PowXN.java
@@ -14,7 +14,7 @@ class PowXN {
         return helper(x, n);
     }
 
-    private double helper(double x, long n) {
+    private double helper(double x, int n) {
         if (n == 0) {
             return 1;
         }
